Time: 5 ms (72.81%), Space: 42.8 MB (58.69%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -4,11 +4,11 @@ class Solution {
             return false;
         }
 
-        long reversed = 0;
-        long temp = x;
+        int reversed = 0;
+        int temp = x;
 
         while (temp != 0) {
-            int digit = (int) (temp % 10);
+            int digit =  (temp % 10);
             reversed = reversed * 10 + digit;
             temp /= 10;
         }
